requirements.txt
# ClipPilot Requirements

# CAMEL-AI Framework
camel-ai>=0.1.0

# MCP Server Dependencies
mcp-python-sdk>=1.0.0

# YouTube MCP Server
# Note: This will be installed via NPX at runtime

# YouTube API
google-api-python-client>=2.100.0
youtube-transcript-api>=0.6.1

# LLM Integration
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.4.0
numpy>=1.24.0
pandas>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.9.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.0
